Feature/ButtonUpToTop
Created component ButtonUpToTop, gave so styles to it. Added this button for AddDrinkPage, DrinksPage, FavoriteDrinksPage. For the others we can add later when page will be done fully.

Added some styles for this button, but can be changed for some improvements.
I took this icon from lucide-react.
Installed new package lucide-react, so everybody should write npm-install in terminal.